Possible multi item bug fix
When 2 items would arrive very close to each other it would drop both items in the first bin and after move to next. Changed the feature that dequeed the first item after it arrived at the 0 different point and made it into a simple delay. Haven't tested, but I think it should solve/improve the issues.